py_experiment/pyproject.toml
[project]
name = "py-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "miniaudio>=1.61",
    "numpy>=2.2.3",
    "pygame-ce>=2.5.3",
    "soundcard>=0.4.3",
    "sounddevice>=0.5.1",
]
